+++ frontend/src/components/teams/InterviewTeam.tsx
@@ -8,7 +8,7 @@ import { GITHUB_AVATAR_SIZE_LIST } from 'constants/constants';
 
 import FlexBox from 'components/@commons/FlexBox';
 import Image from 'components/@commons/Image';
-import { InterviewTeamType, ParticipantType } from 'types/team';
+import { InterviewTeamType } from 'types/team';
 import { convertDateAndTime, teamGetUriBuilder } from 'utils/util';
 
 const InterviewTeam = ({ team }: InterviewTeamsProp) => {
