+++ frontend/src/hooks/preQuestion/usePreQuestion.ts
@@ -6,7 +6,7 @@ import axios, { AxiosError, AxiosResponse } from 'axios';
 import { useMutation, useQuery } from '@tanstack/react-query';
 import { Editor } from '@toast-ui/react-editor';
 
-import useSnackbar from 'hooks/useSnackbar';
+import useSnackbar from 'hooks/utils/useSnackbar';
 
 import { MESSAGE } from 'constants/constants';
