+++ frontend/src/pages/levellogs/LevellogAdd.tsx
@@ -2,8 +2,8 @@ import { useParams } from 'react-router-dom';
 
 import styled from 'styled-components';
 
-import useLevellog from 'hooks/useLevellog';
-import useSnackbar from 'hooks/useSnackbar';
+import useLevellog from 'hooks/levellog/useLevellog';
+import useSnackbar from 'hooks/utils/useSnackbar';
 
 import { MESSAGE } from 'constants/constants';
