+++ frontend/src/pages/preQuestion/PreQuestionAdd.tsx
@@ -3,9 +3,9 @@ import { useNavigate, useParams } from 'react-router-dom';
 
 import styled from 'styled-components';
 
-import useLevellog from 'hooks/useLevellog';
-import usePreQuestion from 'hooks/usePreQuestion';
-import useSnackbar from 'hooks/useSnackbar';
+import useLevellog from 'hooks/levellog/useLevellog';
+import usePreQuestion from 'hooks/preQuestion/usePreQuestion';
+import useSnackbar from 'hooks/utils/useSnackbar';
 
 import { MESSAGE, ROUTES_PATH } from 'constants/constants';
