+++ frontend/src/pages/teams/InterviewTeamAdd.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 
 import styled from 'styled-components';
 
-import useTeamPage from 'hooks/useTeamPage';
+import useTeamPage from 'hooks/team/useTeamPage';
 
 import TeamForm from 'components/teams/TeamForm';
