+++ frontend/src/pages/teams/InterviewTeams.tsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 
 import styled from 'styled-components';
 
-import useTeams from 'hooks/useTeams';
+import useTeams from 'hooks/team/useTeams';
 import useUser from 'hooks/useUser';
 
 import { Exception } from 'pages/status';
